DECA-217: Removing camera capture and calibration from UI
Removing most of the camera capture and calibration work flow from the UI. There is still the mock book that contains a reference to camera capture. This portion should be renamed along with changes to the server to create the proper mock structure. Should be along in another commit shortly.